LayoutTests/fast/mediastream/mediastreamtrack-video-frameRate-increasing.html

<!DOCTYPE html>
<html>
<head>
    <meta charset="utf-8">
    <title>Adapt camera track framerate</title>
    <script src="../../resources/testharness.js"></script>
    <script src="../../resources/testharnessreport.js"></script>
    <script src ="../../webrtc/routines.js"></script>
</head>
<body>
    <video autoplay id="video"></video>
    <script>
promise_test(async (t) => {
    const stream1 = await navigator.mediaDevices.getUserMedia({ video: { width: 160, frameRate : 5 } });
    const stream2 = await navigator.mediaDevices.getUserMedia({ video: { width: 720, frameRate : 30 } });

    // We adapt frame rate after 1 second of samples.
    await new Promise(resolve => setTimeout(resolve, 1000)); 

    if (!stream1.getVideoTracks()[0].muted) {
        const frameRate = await computeFrameRate(stream1, video);
        assert_greater_than(frameRate, 1, "stream1 frame rate above 1");
        assert_less_than(frameRate, 20, "stream1 frame rate below 20");
    }

    const frameRate = await computeFrameRate(stream2, video);
    assert_greater_than(frameRate, 10, "stream2 frame rate above 10");
    assert_less_than(frameRate, 60, "stream2 frame rate below 60");
}, "getUserMedia with increasing frame rates");
    </script>
</body>
</html>
